+++ docs/changelog.rst
@@ -7,7 +7,7 @@ Changelog
 **Added**
 
 - ``st`` as an alias to the ``schemathesis`` command line entrypoint.
-- ``st auth login`` to authenticate with Schemathesis.io.
+- ``st auth login`` / ``st auth logout`` to authenticate with Schemathesis.io.
 
 **Changed**